Extract shot server into standalone process, shared utils into utils.js
- shot-server.js: independent process on port 3031 w/ own browser
- utils.js: detectChallengePage, waitForChallengeBypass, dismissModals, initAdblocker
- daemon.js: rm /shot, /shot-multi, sharp, static serving
- sharp moved to optionalDependencies
