+++ daemon.js
@@ -1,74 +1,15 @@
 const express = require('express');
 const { chromium } = require('patchright');
-// Patchright has stealth features built-in
 const fs = require('fs');
 const os = require('os');
 const path = require('path');
-const sharp = require('sharp');
+const { initAdblocker } = require('./utils');
 
 // Optional ad blocking (off by default)
 // Enable with: BR_ADBLOCK=true br start
 // Base level: BR_ADBLOCK_BASE=none|adsandtrackers|full|ads
 // Additional lists: BR_ADBLOCK_LISTS=https://url1.txt,https://url2.txt
 let adblocker = null;
-
-async function initAdblocker() {
-  const { PlaywrightBlocker } = require('@ghostery/adblocker-playwright');
-  const fetch = require('cross-fetch');
-
-  const base = process.env.BR_ADBLOCK_BASE || 'adsandtrackers';
-  const additionalLists = process.env.BR_ADBLOCK_LISTS;
-
-  // Get base blocker
-  let blocker;
-  switch (base) {
-    case 'none':
-      blocker = PlaywrightBlocker.empty();
-      console.log('Ad blocking enabled (no base filters)');
-      break;
-    case 'full':
-      blocker = await PlaywrightBlocker.fromPrebuiltFull(fetch);
-      console.log('Ad blocking enabled (full: ads + tracking + annoyances + cookies)');
-      break;
-    case 'ads':
-      blocker = await PlaywrightBlocker.fromPrebuiltAdsOnly(fetch);
-      console.log('Ad blocking enabled (ads only)');
-      break;
-    case 'adsandtrackers':
-    default:
-      blocker = await PlaywrightBlocker.fromPrebuiltAdsAndTracking(fetch);
-      console.log('Ad blocking enabled (ads + tracking)');
-      break;
-  }
-
-  // Add additional lists if specified
-  if (additionalLists) {
-    const customLists = additionalLists.split(',').map(s => s.trim());
-
-    for (const listPath of customLists) {
-      let listContent;
-      if (listPath.startsWith('http://') || listPath.startsWith('https://')) {
-        const response = await fetch(listPath);
-        listContent = await response.text();
-      } else {
-        listContent = fs.readFileSync(listPath, 'utf8');
-      }
-
-      const filters = listContent
-        .split('\n')
-        .map(line => line.trim())
-        .filter(line => line.length > 0 && !line.startsWith('!'));
-
-      console.log(`Loaded ${listContent.split('\n').length} lines (${filters.length} active rules) from ${listPath}`);
-
-      blocker.updateFromDiff({ added: filters });
-
-      console.log(`Successfully applied custom list ${listPath}`);
-    }
-  }
-
-  adblocker = blocker;
-}
 
 let lastIdToXPath = {}; // Global variable to store the last idToXPath mapping
 const secrets = new Set();
@@ -107,90 +48,6 @@ async function attachConsoleListeners(page, getTabIndex) {
   });
 }
 
-async function detectChallengePage(page) {
-  try {
-    return await page.evaluate(() => {
-      // Cloudflare
-      if (document.title === 'Just a moment...' ||
-          window._cf_chl_opt ||
-          document.querySelector('script[src*="/cdn-cgi/challenge-platform/"]') ||
-          (document.querySelector('meta[http-equiv="refresh"]') && document.title.includes('Just a moment'))) {
-        return 'cloudflare';
-      }
-
-      // SiteGround
-      if (document.title === 'Robot Challenge Screen' ||
-          window.sgchallenge ||
-          Array.from(document.querySelectorAll('script')).some(script =>
-            script.textContent.includes('sgchallenge'))) {
-        return 'siteground';
-      }
-
-      return false;
-    });
-  } catch (err) {
-    return false;
-  }
-}
-
-async function waitForChallengeBypass(page, maxWaitSeconds = 8) {
-  const startTime = Date.now();
-
-  while (Date.now() - startTime < maxWaitSeconds * 1000) {
-    const challenge = await detectChallengePage(page);
-    if (!challenge) {
-      return true;
-    }
-    await page.waitForTimeout(100);
-  }
-
-  return false;
-}
-
-// Automatically dismisses modals and other UI elements
-async function dismissModals(page) {
-
-  // blind fire an Escape keypress
-  await page.keyboard.press('Escape');
-
-  // look for close buttons to press
-  const selectors = [
-    '[data-dismiss="modal"]',  // bootstrap
-    '[aria-label="Close dialog"]',
-    '[aria-label="Close"]',
-    '[aria-label="button.close"]',
-    '[aria-modal="true"] [aria-label="Close"]',
-    '[aria-modal="true"] [title="Close"]',
-    '[aria-modal="true"] [data-action="close"]',
-    '.popup .close-button',
-    '.modal .close',
-    'a.close-popup',
-    '[role="dialog"] .close-btn',
-    '[role="dialog"] .close-button',
-    '[role="dialog"] .close',
-    '[role="dialog"] [aria-label="Close"]',
-    'button[data-testid="close-welcome-modal"]',
-    'button.spu-close-popup',
-    '#campaign_modal_wrapper #continue_to_site',
-    '.close-footer-btn'
-  ].join(', ');
-  const maxWaitTime = 2500;
-  const startTime = Date.now();
-
-  while (Date.now() - startTime < maxWaitTime) {
-    const closeButton = await page.$(selectors);
-    if (!closeButton) {
-      break;
-    }
-
-    if (await closeButton.isVisible().catch(() => false)) {
-      await closeButton.click().catch(() => {});
-    }
-
-    await page.waitForTimeout(500);
-  }
-}
-
 function record(action, args = {}) {
   history.push({ action, args, timestamp: new Date().toISOString() });
 }
@@ -206,7 +63,7 @@ const tmpUserDataDir = path.join(os.tmpdir(), `br_user_data_${Date.now()}`);
 (async () => {
   // Initialize adblocker if enabled
   if (process.env.BR_ADBLOCK === 'true') {
-    await initAdblocker();
+    adblocker = await initAdblocker();
   }
   // Set viewport size for headless mode
   let viewport = null;
@@ -284,9 +141,6 @@ const tmpUserDataDir = path.join(os.tmpdir(), `br_user_data_${Date.now()}`);
   });
 
   app.use(express.json());
-
-  // Serve static files
-  app.use(express.static(path.join(__dirname, 'public')));
 
   app.get('/health', (req, res) => {
     res.send('ok');
@@ -763,225 +617,6 @@ Try using --selector to specify the search input explicitly.`);
 
     } catch (err) {
       res.status(500).send(`Error extracting text: ${err.message}`);
-    }
-  });
-
-  app.post('/shot', async (req, res) => {
-    let page;
-    try {
-      const {
-        url,
-        width,
-        height,
-        waitTime,
-        output_width,
-        output_format,
-        output_quality
-      } = req.body;
-
-      if (!url) {
-        return res.status(400).json({ error: 'missing url parameter' });
-      }
-
-      // Prepend https:// if no protocol is specified
-      let processedUrl = url;
-      if (!url.startsWith('http://') && !url.startsWith('https://')) {
-        processedUrl = `https://${url}`;
-      }
-
-      // Create a new page for the screenshot
-      page = await context.newPage();
-      if (adblocker) await adblocker.enableBlockingInPage(page);
-
-      // Set viewport size if width and height are provided
-      if (width || height) {
-        const viewportWidth = width ? parseInt(width, 10) : 1280;
-        const viewportHeight = height ? parseInt(height, 10) : 720;
-        await page.setViewportSize({ width: viewportWidth, height: viewportHeight });
-      }
-
-      // Navigate to the URL and wait for it to load
-      await page.goto(processedUrl, { timeout: 20000 });
-
-      // Wait for challenge pages to complete before screenshotting
-      await waitForChallengeBypass(page);
-
-      // Add style to hide scrollbars for cleaner screenshots
-      await page.addStyleTag({
-        content: `
-          /* WebKit scrollbar hiding (Chrome, Safari) */
-          ::-webkit-scrollbar { display: none; }
-
-          /* Firefox */
-          html { scrollbar-width: none; }
-
-          /* IE and Edge */
-          html { -ms-overflow-style: none; }
-        `
-      });
-
-      // Wait additional time if specified (default: 1000ms)
-      const additionalWait = waitTime ? parseInt(waitTime, 10) : 1000;
-      await page.waitForTimeout(additionalWait);
-
-      // dismiss any modals not caught by adblock
-      await dismissModals(page);
-
-      // Take screenshot to buffer
-      const isFullPage = !height; // Full page if height not provided
-      const screenshotBuffer = await page.screenshot({
-        fullPage: isFullPage,
-        type: 'png'
-      });
-
-      record('api-shot', { url, width, height, fullPage: isFullPage, waitTime: additionalWait });
-
-      // Apply output transformations if specified
-      let processedBuffer = screenshotBuffer;
-      const format = output_format || 'png';
-      const quality = output_quality ? parseInt(output_quality, 10) : 80;
-
-      // Create sharp processor
-      let processor = sharp(screenshotBuffer);
-
-      // Resize if output_width is specified
-      if (output_width) {
-        const outputWidth = parseInt(output_width, 10);
-        processor = processor.resize(outputWidth, null, {
-          withoutEnlargement: true
-        });
-      }
-
-      // Convert to the requested format
-      if (format === 'webp') {
-        processedBuffer = await processor.webp({ quality }).toBuffer();
-        res.setHeader('Content-Type', 'image/webp');
-      } else if (format === 'jpeg' || format === 'jpg') {
-        processedBuffer = await processor.jpeg({ quality }).toBuffer();
-        res.setHeader('Content-Type', 'image/jpeg');
-      } else {
-        processedBuffer = await processor.png().toBuffer();
-        res.setHeader('Content-Type', 'image/png');
-      }
-
-      res.setHeader('Content-Length', processedBuffer.length);
-      res.send(processedBuffer);
-
-    } catch (err) {
-      res.status(500).json({ error: `Error capturing screenshot: ${err.message}` });
-    } finally {
-      // Close the page if it was created
-      if (page) {
-        await page.close();
-      }
-    }
-  });
-
-  // Multiple screenshots with different dimensions from single page load
-  app.post('/shot-multi', async (req, res) => {
-    let page;
-    try {
-      const {
-        url,
-        width,
-        waitTime,
-        outputs,
-        output_format,
-        output_quality
-      } = req.body;
-
-      if (!url) {
-        return res.status(400).json({ error: 'missing url parameter' });
-      }
-      if (!outputs || !Array.isArray(outputs) || outputs.length === 0) {
-        return res.status(400).json({ error: 'outputs must be a non-empty array' });
-      }
-
-      // Prepend https:// if no protocol is specified
-      let processedUrl = url;
-      if (!url.startsWith('http://') && !url.startsWith('https://')) {
-        processedUrl = `https://${url}`;
-      }
-
-      page = await context.newPage();
-      if (adblocker) await adblocker.enableBlockingInPage(page);
-
-      const maxWidth = width ? parseInt(width, 10) : 1280;
-      const maxPixelHeight = outputs.reduce((max, o) => Math.max(max, parseInt(o.height) || 0), 0) || null;
-      await page.setViewportSize({ width: maxWidth, height: maxPixelHeight || 720 });
-
-      await page.goto(processedUrl, { timeout: 20000 });
-      await waitForChallengeBypass(page);
-
-      await page.addStyleTag({
-        content: `
-          ::-webkit-scrollbar { display: none; }
-          html { scrollbar-width: none; }
-          html { -ms-overflow-style: none; }
-        `
-      });
-
-      const additionalWait = waitTime ? parseInt(waitTime, 10) : 1000;
-      await page.waitForTimeout(additionalWait);
-      await dismissModals(page);
-
-      const format = output_format || 'png';
-      const quality = output_quality ? parseInt(output_quality, 10) : 80;
-
-      const fullPage = !maxPixelHeight;
-      const screenshotBuffer = await page.screenshot({ fullPage, type: 'png' });
-      const metadata = await sharp(screenshotBuffer).metadata();
-
-      const images = await Promise.all(outputs.map(async (output) => {
-        const outputHeight = output.height ? parseInt(output.height, 10) : null;
-        const outputWidth = output.output_width ? parseInt(output.output_width, 10) : null;
-        const outputFormat = output.output_format || format;
-        const outputQuality = output.output_quality ? parseInt(output.output_quality, 10) : quality;
-
-        let processor = sharp(screenshotBuffer);
-
-        // Crop to height from top
-        if (outputHeight && metadata.height > outputHeight) {
-          processor = processor.extract({ left: 0, top: 0, width: metadata.width, height: outputHeight });
-        }
-
-        // Resize to output_width
-        if (outputWidth) {
-          processor = processor.resize(outputWidth, null, { withoutEnlargement: true });
-        }
-
-        let buffer;
-        let contentType;
-        if (outputFormat === 'webp') {
-          buffer = await processor.webp({ quality: outputQuality }).toBuffer();
-          contentType = 'image/webp';
-        } else if (outputFormat === 'jpeg' || outputFormat === 'jpg') {
-          buffer = await processor.jpeg({ quality: outputQuality }).toBuffer();
-          contentType = 'image/jpeg';
-        } else {
-          buffer = await processor.toBuffer();
-          contentType = 'image/png';
-        }
-
-        const finalMetadata = await sharp(buffer).metadata();
-        return {
-          data: buffer.toString('base64'),
-          content_type: contentType,
-          height: finalMetadata.height,
-          width: finalMetadata.width
-        };
-      }));
-
-      record('api-shot-multi', { url, width, outputs, fullPage });
-
-      res.json({ images });
-
-    } catch (err) {
-      res.status(500).json({ error: `Error capturing screenshots: ${err.message}` });
-    } finally {
-      if (page) {
-        await page.close();
-      }
     }
   });
