+++ package.json
@@ -10,7 +10,9 @@
     "commander": "^11.1.0",
     "cross-fetch": "^4.1.0",
     "express": "^4.18.2",
-    "patchright": "^1.56.1",
+    "patchright": "^1.56.1"
+  },
+  "optionalDependencies": {
     "sharp": "^0.34.5"
   }
 }
